+++ apps/menu-bar/macos/ExpoMenuBar-macOS/SwifterWrapper.swift
@@ -3,7 +3,7 @@ import Swifter
 import Dispatch
 
 private let PORTS = [35783, 47909, 44171, 50799]
-private let WHITELISTED_DOMAINS = ["expo.dev", "expo.test", "exp.host"]
+private let WHITELISTED_DOMAINS = ["expo.dev", "expo.test", "exp.host", "localhost"]
 
 @objc class SwifterWrapper: NSObject {
   let server = HttpServer()
@@ -98,7 +98,7 @@ private let WHITELISTED_DOMAINS = ["expo.dev", "expo.test", "exp.host"]
       }
     }
 
-    if !hostName.contains(".") {
+    if !hostName.contains(".") && originUrl.pathComponents.count > 1 {
       hostName = originUrl.pathComponents[1]
     }