+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Upgrade to `expo` SDK 54 and react-native `0.81`. ([#254](https://github.com/expo/orbit/pull/254) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 - Optimize code for React 19. ([#307](https://github.com/expo/orbit/pull/307) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 - Upgrade `react-native-svg` to 15.15.2. ([#308](https://github.com/expo/orbit/pull/308) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+- Add localhost to local server whitelisted domains. ([#317](https://github.com/expo/orbit/pull/317) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 
 ## 2.2.0 — 2025-11-11